FocusEval_Web_GPT/backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow 